Moved DateTime class to time package
