+++ src/com/jboby93/markovbot/App.java
@@ -11,6 +11,7 @@ import java.util.Scanner;
 
 import db.DBSearchResult;
 import logging.Logger;
+import time.DateTime;
 
 public class App {
 	public static final String name = "MarkovBot 98";
